pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Eduard Schwarzkopf <48969167+EduardSchwarzkopf@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
arel = "0.3.0"
fastapi = "0.109.2"
fastapi-mail = "1.4.1"
fastapi-users = "12.1.3"
fastapi-users-db-sqlalchemy = "6.0.1"
httpx = "0.26.0"
pydantic = "^2.6.1"
pytest = "8.0.0"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
sqlalchemy = "2.0.27"
starlette = "0.36.3"
starlette-wtf = "0.4.3"
uvicorn = "0.27.1"
wtforms = "3.1.2"
setuptools = "^69.1.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
bcrypt = "^4.1.2"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
follow_imports = "skip"
exclude = "alembic"